Remove Performance/RangeInclude cop
This cop was introduced[1] to avoid uses of `Range#include?` when
`Range#cover?` would suffice. The justification was that the former
iterated over the entire range, while the latter did a bounds check on
the start and end of the range.

However, since at least ruby 2.2, it hasn't been the case that
`Range#include?` iterates over its contents; bounds checking is used
where appropriate.

This is borne out by benchmarking - in the latest patch versions of
rubies 2.2 to 2.5 at the time of writing, `Range#cover?` was not
detectably faster than `Range#include?` given a 10k integer range,
testing for a value at its midpoint.

Since there doesn't seem to be any current performance justification for
this cop, nor a particular stylistic reason for preferring `#cover?` to
the more familiar `#include?`, this change removes the cop.

[1] be2f41fd2