Extract obsolete cop and parameter config to YAML files
The configuration for obsolete cops and parameters was previously
embedded in the Config class as increasingly lengthy hash and array
literals.

Since managing the multiline strings was something of a hassle, this
extracts the config to two YAML files.

In order to parse these where required without duplication or
overloading the ConfigLoader class, I've extracted some of the latter's
safe YAML-handling methods to a helper class. The ConfigLoader wasn't
directly usable as it relates to parsing the main .rubocop.yml
configuration, and asserts certain properties of the config that don't
apply to these new, simpler (non-customisable) config files.